employee_management/requirements.txt
Flask==2.0.2
kafka-python==2.0.2